+++ src/webhook/types/ping.rs
@@ -7,7 +7,7 @@ use super::common::{Repository, User};
 #[derive(Debug, Deserialize)]
 pub struct PingEvent {
     pub zen: String,
-    pub hook_id: u32,
+    pub hook_id: u64,
     pub repository: Repository,
     pub sender: User,
 }
